backend/requirements.txt
fastapi==0.115.5
email-validator==2.2.0
uvicorn==0.32.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.17
